updated in 32/05/2023

+++ tarefa03/tarefa_03.c
@@ -9,7 +9,7 @@
 //structed data 
 struct bioData {
     float altura;
-    int peso;
+    float peso;
 };
 
 typedef struct bioData BioData;
